Move StateT instances into one place
